Move imap functionality into concern
