+++ app/controllers/mails_controller.rb
@@ -9,7 +9,7 @@ require 'net/imap'
 
 class MailsController < ApplicationController
 
-  after_action :disconnect
+  include Imap
 
   skip_authorization_check
 
@@ -60,82 +60,25 @@ class MailsController < ApplicationController
     params[:mailbox]
   end
 
-  def imap
-    if @imap.present?
-      return @imap
-    end
-
-    imap = Net::IMAP.new('imap.gmail.com', 993, true)
-    imap.login(email, password)
-    @imap ||= imap
-  end
-
-  def disconnect
-    imap.close
-    imap.disconnect
-  end
-
   def mails
     @mails ||= { inbox_mails: inbox_mails, spam_mails: spam_mails, failed_mails: failed_mails }
   end
 
   def failed_mails
-    @failed_mails ||= mailbox_mails(FAILED)
+    @failed_mails ||= fetch_all_from_mailbox(FAILED.id)
   end
 
   def spam_mails
-    @spam_mails ||= mailbox_mails(SPAM)
+    @spam_mails ||= fetch_all_from_mailbox(SPAM.id)
   end
 
   def inbox_mails
-    @inbox_mails ||= mailbox_mails(INBOX)
-  end
-
-  def mailbox_mails(mailbox = INBOX)
-    imap.select(mailbox.id)
-
-    num_messages = imap.status(mailbox.id, ['MESSAGES'])['MESSAGES']
-    mails = if num_messages.positive?
-              @imap.fetch(1..num_messages, attributes) || []
-            else
-              []
-            end
-
-    mails.map { |m| CatchAllMail.new(imap_fetch_data=m, mailbox=mailbox.id) }
-
+    @inbox_mails ||= fetch_all_from_mailbox(INBOX.id)
   end
 
   def mail
     @mail ||= fetch_by_uid(param_uid, param_mailbox)
   end
 
-  def fetch_by_uid(uid, mailbox_id = INBOX.id)
-    imap.select(mailbox_id)
-    fetch_data = imap.uid_fetch(uid, attributes)[0]
-    CatchAllMail.new(fetch_data, mailbox_id)
-  end
 
-  def move_by_uid(uid, from_mailbox_id, to_mailbox_id)
-    imap.select(from_mailbox_id)
-    imap.uid_move(uid, to_mailbox_id)
-  end
-
-  def delete_by_uid(uid, mailbox_id)
-    imap.select(mailbox_id)
-    # imap.uid_copy(uid, 'TRASH')
-    imap.uid_store(uid, '+FLAGS', [:Deleted])
-    imap.expunge
-  end
-
-  def email
-    ENV.fetch('USER_EMAIL', 'test.imap.hitobito@gmail.com')
-  end
-
-  def password
-    ENV.fetch('USER_PASSWORD', 'test.imap')
-  end
-
-  def attributes
-    %w(ENVELOPE UID BODYSTRUCTURE BODY[TEXT])
-  end
 end
